+++ caffe/src/main/java/org/bytedeco/javacpp/caffe.java
@@ -22551,13 +22551,13 @@ public static final String HDF5_DATA_LABEL_NAME = "label";
 /**
  * \brief CuDNN acceleration of ReLULayer.
  */
-@Platform({"linux-x86_64", "macosx-x86_64"}) @Name("caffe::CuDNNReLULayer<float>") @NoOffset public static class FloatCuDNNReLULayer extends FloatReLULayer {
+@Platform(value = {"linux-x86_64", "macosx-x86_64"}, extension = "-gpu") @Name("caffe::CuDNNReLULayer<float>") @NoOffset public static class FloatCuDNNReLULayer extends FloatReLULayer {
     static { Loader.load(); }
     /** Pointer cast constructor. Invokes {@link Pointer#Pointer(Pointer)}. */
     public FloatCuDNNReLULayer(Pointer p) { super(p); }
 
   public FloatCuDNNReLULayer(@Const @ByRef LayerParameter param) { super((Pointer)null); allocate(param); }
-  private native @Platform({"linux-x86_64", "macosx-x86_64"}) void allocate(@Const @ByRef LayerParameter param);
+  private native @Platform(value = {"linux-x86_64", "macosx-x86_64"}, extension = "-gpu") void allocate(@Const @ByRef LayerParameter param);
   @Virtual public native void LayerSetUp(@Const @ByRef FloatBlobVector bottom,
         @Const @ByRef FloatBlobVector top);
   @Virtual public native void Reshape(@Const @ByRef FloatBlobVector bottom,
@@ -22567,13 +22567,13 @@ public static final String HDF5_DATA_LABEL_NAME = "label";
   @Virtual protected native void Backward_gpu(@Const @ByRef FloatBlobVector top,
         @Const @ByRef BoolVector propagate_down, @Const @ByRef FloatBlobVector bottom);
 }
-@Platform({"linux-x86_64", "macosx-x86_64"}) @Name("caffe::CuDNNReLULayer<double>") @NoOffset public static class DoubleCuDNNReLULayer extends DoubleReLULayer {
+@Platform(value = {"linux-x86_64", "macosx-x86_64"}, extension = "-gpu") @Name("caffe::CuDNNReLULayer<double>") @NoOffset public static class DoubleCuDNNReLULayer extends DoubleReLULayer {
     static { Loader.load(); }
     /** Pointer cast constructor. Invokes {@link Pointer#Pointer(Pointer)}. */
     public DoubleCuDNNReLULayer(Pointer p) { super(p); }
 
   public DoubleCuDNNReLULayer(@Const @ByRef LayerParameter param) { super((Pointer)null); allocate(param); }
-  private native @Platform({"linux-x86_64", "macosx-x86_64"}) void allocate(@Const @ByRef LayerParameter param);
+  private native @Platform(value = {"linux-x86_64", "macosx-x86_64"}, extension = "-gpu") void allocate(@Const @ByRef LayerParameter param);
   @Virtual public native void LayerSetUp(@Const @ByRef DoubleBlobVector bottom,
         @Const @ByRef DoubleBlobVector top);
   @Virtual public native void Reshape(@Const @ByRef DoubleBlobVector bottom,
@@ -22671,13 +22671,13 @@ public static final String HDF5_DATA_LABEL_NAME = "label";
 /**
  * \brief CuDNN acceleration of SigmoidLayer.
  */
-@Platform({"linux-x86_64", "macosx-x86_64"}) @Name("caffe::CuDNNSigmoidLayer<float>") @NoOffset public static class FloatCuDNNSigmoidLayer extends FloatSigmoidLayer {
+@Platform(value = {"linux-x86_64", "macosx-x86_64"}, extension = "-gpu") @Name("caffe::CuDNNSigmoidLayer<float>") @NoOffset public static class FloatCuDNNSigmoidLayer extends FloatSigmoidLayer {
     static { Loader.load(); }
     /** Pointer cast constructor. Invokes {@link Pointer#Pointer(Pointer)}. */
     public FloatCuDNNSigmoidLayer(Pointer p) { super(p); }
 
   public FloatCuDNNSigmoidLayer(@Const @ByRef LayerParameter param) { super((Pointer)null); allocate(param); }
-  private native @Platform({"linux-x86_64", "macosx-x86_64"}) void allocate(@Const @ByRef LayerParameter param);
+  private native @Platform(value = {"linux-x86_64", "macosx-x86_64"}, extension = "-gpu") void allocate(@Const @ByRef LayerParameter param);
   @Virtual public native void LayerSetUp(@Const @ByRef FloatBlobVector bottom,
         @Const @ByRef FloatBlobVector top);
   @Virtual public native void Reshape(@Const @ByRef FloatBlobVector bottom,
@@ -22687,13 +22687,13 @@ public static final String HDF5_DATA_LABEL_NAME = "label";
   @Virtual protected native void Backward_gpu(@Const @ByRef FloatBlobVector top,
         @Const @ByRef BoolVector propagate_down, @Const @ByRef FloatBlobVector bottom);
 }
-@Platform({"linux-x86_64", "macosx-x86_64"}) @Name("caffe::CuDNNSigmoidLayer<double>") @NoOffset public static class DoubleCuDNNSigmoidLayer extends DoubleSigmoidLayer {
+@Platform(value = {"linux-x86_64", "macosx-x86_64"}, extension = "-gpu") @Name("caffe::CuDNNSigmoidLayer<double>") @NoOffset public static class DoubleCuDNNSigmoidLayer extends DoubleSigmoidLayer {
     static { Loader.load(); }
     /** Pointer cast constructor. Invokes {@link Pointer#Pointer(Pointer)}. */
     public DoubleCuDNNSigmoidLayer(Pointer p) { super(p); }
 
   public DoubleCuDNNSigmoidLayer(@Const @ByRef LayerParameter param) { super((Pointer)null); allocate(param); }
-  private native @Platform({"linux-x86_64", "macosx-x86_64"}) void allocate(@Const @ByRef LayerParameter param);
+  private native @Platform(value = {"linux-x86_64", "macosx-x86_64"}, extension = "-gpu") void allocate(@Const @ByRef LayerParameter param);
   @Virtual public native void LayerSetUp(@Const @ByRef DoubleBlobVector bottom,
         @Const @ByRef DoubleBlobVector top);
   @Virtual public native void Reshape(@Const @ByRef DoubleBlobVector bottom,
@@ -22791,13 +22791,13 @@ public static final String HDF5_DATA_LABEL_NAME = "label";
 /**
  * \brief CuDNN acceleration of TanHLayer.
  */
-@Platform({"linux-x86_64", "macosx-x86_64"}) @Name("caffe::CuDNNTanHLayer<float>") @NoOffset public static class FloatCuDNNTanHLayer extends FloatTanHLayer {
+@Platform(value = {"linux-x86_64", "macosx-x86_64"}, extension = "-gpu") @Name("caffe::CuDNNTanHLayer<float>") @NoOffset public static class FloatCuDNNTanHLayer extends FloatTanHLayer {
     static { Loader.load(); }
     /** Pointer cast constructor. Invokes {@link Pointer#Pointer(Pointer)}. */
     public FloatCuDNNTanHLayer(Pointer p) { super(p); }
 
   public FloatCuDNNTanHLayer(@Const @ByRef LayerParameter param) { super((Pointer)null); allocate(param); }
-  private native @Platform({"linux-x86_64", "macosx-x86_64"}) void allocate(@Const @ByRef LayerParameter param);
+  private native @Platform(value = {"linux-x86_64", "macosx-x86_64"}, extension = "-gpu") void allocate(@Const @ByRef LayerParameter param);
   @Virtual public native void LayerSetUp(@Const @ByRef FloatBlobVector bottom,
         @Const @ByRef FloatBlobVector top);
   @Virtual public native void Reshape(@Const @ByRef FloatBlobVector bottom,
@@ -22807,13 +22807,13 @@ public static final String HDF5_DATA_LABEL_NAME = "label";
   @Virtual protected native void Backward_gpu(@Const @ByRef FloatBlobVector top,
         @Const @ByRef BoolVector propagate_down, @Const @ByRef FloatBlobVector bottom);
 }
-@Platform({"linux-x86_64", "macosx-x86_64"}) @Name("caffe::CuDNNTanHLayer<double>") @NoOffset public static class DoubleCuDNNTanHLayer extends DoubleTanHLayer {
+@Platform(value = {"linux-x86_64", "macosx-x86_64"}, extension = "-gpu") @Name("caffe::CuDNNTanHLayer<double>") @NoOffset public static class DoubleCuDNNTanHLayer extends DoubleTanHLayer {
     static { Loader.load(); }
     /** Pointer cast constructor. Invokes {@link Pointer#Pointer(Pointer)}. */
     public DoubleCuDNNTanHLayer(Pointer p) { super(p); }
 
   public DoubleCuDNNTanHLayer(@Const @ByRef LayerParameter param) { super((Pointer)null); allocate(param); }
-  private native @Platform({"linux-x86_64", "macosx-x86_64"}) void allocate(@Const @ByRef LayerParameter param);
+  private native @Platform(value = {"linux-x86_64", "macosx-x86_64"}, extension = "-gpu") void allocate(@Const @ByRef LayerParameter param);
   @Virtual public native void LayerSetUp(@Const @ByRef DoubleBlobVector bottom,
         @Const @ByRef DoubleBlobVector top);
   @Virtual public native void Reshape(@Const @ByRef DoubleBlobVector bottom,
@@ -24020,13 +24020,13 @@ public static final String HDF5_DATA_LABEL_NAME = "label";
  * \brief cuDNN implementation of SoftmaxLayer.
  *        Fallback to SoftmaxLayer for CPU mode.
  */
-@Platform({"linux-x86_64", "macosx-x86_64"}) @Name("caffe::CuDNNSoftmaxLayer<float>") @NoOffset public static class FloatCuDNNSoftmaxLayer extends FloatSoftmaxLayer {
+@Platform(value = {"linux-x86_64", "macosx-x86_64"}, extension = "-gpu") @Name("caffe::CuDNNSoftmaxLayer<float>") @NoOffset public static class FloatCuDNNSoftmaxLayer extends FloatSoftmaxLayer {
     static { Loader.load(); }
     /** Pointer cast constructor. Invokes {@link Pointer#Pointer(Pointer)}. */
     public FloatCuDNNSoftmaxLayer(Pointer p) { super(p); }
 
   public FloatCuDNNSoftmaxLayer(@Const @ByRef LayerParameter param) { super((Pointer)null); allocate(param); }
-  private native @Platform({"linux-x86_64", "macosx-x86_64"}) void allocate(@Const @ByRef LayerParameter param);
+  private native @Platform(value = {"linux-x86_64", "macosx-x86_64"}, extension = "-gpu") void allocate(@Const @ByRef LayerParameter param);
   @Virtual public native void LayerSetUp(@Const @ByRef FloatBlobVector bottom,
         @Const @ByRef FloatBlobVector top);
   @Virtual public native void Reshape(@Const @ByRef FloatBlobVector bottom,
@@ -24036,13 +24036,13 @@ public static final String HDF5_DATA_LABEL_NAME = "label";
   @Virtual protected native void Backward_gpu(@Const @ByRef FloatBlobVector top,
        @Const @ByRef BoolVector propagate_down, @Const @ByRef FloatBlobVector bottom);
 }
-@Platform({"linux-x86_64", "macosx-x86_64"}) @Name("caffe::CuDNNSoftmaxLayer<double>") @NoOffset public static class DoubleCuDNNSoftmaxLayer extends DoubleSoftmaxLayer {
+@Platform(value = {"linux-x86_64", "macosx-x86_64"}, extension = "-gpu") @Name("caffe::CuDNNSoftmaxLayer<double>") @NoOffset public static class DoubleCuDNNSoftmaxLayer extends DoubleSoftmaxLayer {
     static { Loader.load(); }
     /** Pointer cast constructor. Invokes {@link Pointer#Pointer(Pointer)}. */
     public DoubleCuDNNSoftmaxLayer(Pointer p) { super(p); }
 
   public DoubleCuDNNSoftmaxLayer(@Const @ByRef LayerParameter param) { super((Pointer)null); allocate(param); }
-  private native @Platform({"linux-x86_64", "macosx-x86_64"}) void allocate(@Const @ByRef LayerParameter param);
+  private native @Platform(value = {"linux-x86_64", "macosx-x86_64"}, extension = "-gpu") void allocate(@Const @ByRef LayerParameter param);
   @Virtual public native void LayerSetUp(@Const @ByRef DoubleBlobVector bottom,
         @Const @ByRef DoubleBlobVector top);
   @Virtual public native void Reshape(@Const @ByRef DoubleBlobVector bottom,
@@ -25781,13 +25781,13 @@ public static final String HDF5_DATA_LABEL_NAME = "label";
  * but for fully-convolutional models and large inputs the CAFFE engine can be
  * faster as long as it fits in memory.
 */
-@Platform({"linux-x86_64", "macosx-x86_64"}) @Name("caffe::CuDNNConvolutionLayer<float>") @NoOffset public static class FloatCuDNNConvolutionLayer extends FloatConvolutionLayer {
+@Platform(value = {"linux-x86_64", "macosx-x86_64"}, extension = "-gpu") @Name("caffe::CuDNNConvolutionLayer<float>") @NoOffset public static class FloatCuDNNConvolutionLayer extends FloatConvolutionLayer {
     static { Loader.load(); }
     /** Pointer cast constructor. Invokes {@link Pointer#Pointer(Pointer)}. */
     public FloatCuDNNConvolutionLayer(Pointer p) { super(p); }
 
   public FloatCuDNNConvolutionLayer(@Const @ByRef LayerParameter param) { super((Pointer)null); allocate(param); }
-  private native @Platform({"linux-x86_64", "macosx-x86_64"}) void allocate(@Const @ByRef LayerParameter param);
+  private native @Platform(value = {"linux-x86_64", "macosx-x86_64"}, extension = "-gpu") void allocate(@Const @ByRef LayerParameter param);
   @Virtual public native void LayerSetUp(@Const @ByRef FloatBlobVector bottom,
         @Const @ByRef FloatBlobVector top);
   @Virtual public native void Reshape(@Const @ByRef FloatBlobVector bottom,
@@ -25797,13 +25797,13 @@ public static final String HDF5_DATA_LABEL_NAME = "label";
   @Virtual protected native void Backward_gpu(@Const @ByRef FloatBlobVector top,
         @Const @ByRef BoolVector propagate_down, @Const @ByRef FloatBlobVector bottom);
 }
-@Platform({"linux-x86_64", "macosx-x86_64"}) @Name("caffe::CuDNNConvolutionLayer<double>") @NoOffset public static class DoubleCuDNNConvolutionLayer extends DoubleConvolutionLayer {
+@Platform(value = {"linux-x86_64", "macosx-x86_64"}, extension = "-gpu") @Name("caffe::CuDNNConvolutionLayer<double>") @NoOffset public static class DoubleCuDNNConvolutionLayer extends DoubleConvolutionLayer {
     static { Loader.load(); }
     /** Pointer cast constructor. Invokes {@link Pointer#Pointer(Pointer)}. */
     public DoubleCuDNNConvolutionLayer(Pointer p) { super(p); }
 
   public DoubleCuDNNConvolutionLayer(@Const @ByRef LayerParameter param) { super((Pointer)null); allocate(param); }
-  private native @Platform({"linux-x86_64", "macosx-x86_64"}) void allocate(@Const @ByRef LayerParameter param);
+  private native @Platform(value = {"linux-x86_64", "macosx-x86_64"}, extension = "-gpu") void allocate(@Const @ByRef LayerParameter param);
   @Virtual public native void LayerSetUp(@Const @ByRef DoubleBlobVector bottom,
         @Const @ByRef DoubleBlobVector top);
   @Virtual public native void Reshape(@Const @ByRef DoubleBlobVector bottom,
@@ -26007,13 +26007,13 @@ public static final String HDF5_DATA_LABEL_NAME = "label";
 // #include "caffe/layers/lrn_layer.hpp"
 
 // #ifdef USE_CUDNN
-@Platform({"linux-x86_64", "macosx-x86_64"}) @Name("caffe::CuDNNLRNLayer<float>") @NoOffset public static class FloatCuDNNLRNLayer extends FloatLRNLayer {
+@Platform(value = {"linux-x86_64", "macosx-x86_64"}, extension = "-gpu") @Name("caffe::CuDNNLRNLayer<float>") @NoOffset public static class FloatCuDNNLRNLayer extends FloatLRNLayer {
     static { Loader.load(); }
     /** Pointer cast constructor. Invokes {@link Pointer#Pointer(Pointer)}. */
     public FloatCuDNNLRNLayer(Pointer p) { super(p); }
 
   public FloatCuDNNLRNLayer(@Const @ByRef LayerParameter param) { super((Pointer)null); allocate(param); }
-  private native @Platform({"linux-x86_64", "macosx-x86_64"}) void allocate(@Const @ByRef LayerParameter param);
+  private native @Platform(value = {"linux-x86_64", "macosx-x86_64"}, extension = "-gpu") void allocate(@Const @ByRef LayerParameter param);
   @Virtual public native void LayerSetUp(@Const @ByRef FloatBlobVector bottom,
         @Const @ByRef FloatBlobVector top);
   @Virtual public native void Reshape(@Const @ByRef FloatBlobVector bottom,
@@ -26023,13 +26023,13 @@ public static final String HDF5_DATA_LABEL_NAME = "label";
   @Virtual protected native void Backward_gpu(@Const @ByRef FloatBlobVector top,
         @Const @ByRef BoolVector propagate_down, @Const @ByRef FloatBlobVector bottom);
 }
-@Platform({"linux-x86_64", "macosx-x86_64"}) @Name("caffe::CuDNNLRNLayer<double>") @NoOffset public static class DoubleCuDNNLRNLayer extends DoubleLRNLayer {
+@Platform(value = {"linux-x86_64", "macosx-x86_64"}, extension = "-gpu") @Name("caffe::CuDNNLRNLayer<double>") @NoOffset public static class DoubleCuDNNLRNLayer extends DoubleLRNLayer {
     static { Loader.load(); }
     /** Pointer cast constructor. Invokes {@link Pointer#Pointer(Pointer)}. */
     public DoubleCuDNNLRNLayer(Pointer p) { super(p); }
 
   public DoubleCuDNNLRNLayer(@Const @ByRef LayerParameter param) { super((Pointer)null); allocate(param); }
-  private native @Platform({"linux-x86_64", "macosx-x86_64"}) void allocate(@Const @ByRef LayerParameter param);
+  private native @Platform(value = {"linux-x86_64", "macosx-x86_64"}, extension = "-gpu") void allocate(@Const @ByRef LayerParameter param);
   @Virtual public native void LayerSetUp(@Const @ByRef DoubleBlobVector bottom,
         @Const @ByRef DoubleBlobVector top);
   @Virtual public native void Reshape(@Const @ByRef DoubleBlobVector bottom,
@@ -26061,13 +26061,13 @@ public static final String HDF5_DATA_LABEL_NAME = "label";
 // #include "caffe/layers/power_layer.hpp"
 
 // #ifdef USE_CUDNN
-@Platform({"linux-x86_64", "macosx-x86_64"}) @Name("caffe::CuDNNLCNLayer<float>") @NoOffset public static class FloatCuDNNLCNLayer extends FloatLRNLayer {
+@Platform(value = {"linux-x86_64", "macosx-x86_64"}, extension = "-gpu") @Name("caffe::CuDNNLCNLayer<float>") @NoOffset public static class FloatCuDNNLCNLayer extends FloatLRNLayer {
     static { Loader.load(); }
     /** Pointer cast constructor. Invokes {@link Pointer#Pointer(Pointer)}. */
     public FloatCuDNNLCNLayer(Pointer p) { super(p); }
 
   public FloatCuDNNLCNLayer(@Const @ByRef LayerParameter param) { super((Pointer)null); allocate(param); }
-  private native @Platform({"linux-x86_64", "macosx-x86_64"}) void allocate(@Const @ByRef LayerParameter param);
+  private native @Platform(value = {"linux-x86_64", "macosx-x86_64"}, extension = "-gpu") void allocate(@Const @ByRef LayerParameter param);
   @Virtual public native void LayerSetUp(@Const @ByRef FloatBlobVector bottom,
         @Const @ByRef FloatBlobVector top);
   @Virtual public native void Reshape(@Const @ByRef FloatBlobVector bottom,
@@ -26077,13 +26077,13 @@ public static final String HDF5_DATA_LABEL_NAME = "label";
   @Virtual protected native void Backward_gpu(@Const @ByRef FloatBlobVector top,
         @Const @ByRef BoolVector propagate_down, @Const @ByRef FloatBlobVector bottom);
 }
-@Platform({"linux-x86_64", "macosx-x86_64"}) @Name("caffe::CuDNNLCNLayer<double>") @NoOffset public static class DoubleCuDNNLCNLayer extends DoubleLRNLayer {
+@Platform(value = {"linux-x86_64", "macosx-x86_64"}, extension = "-gpu") @Name("caffe::CuDNNLCNLayer<double>") @NoOffset public static class DoubleCuDNNLCNLayer extends DoubleLRNLayer {
     static { Loader.load(); }
     /** Pointer cast constructor. Invokes {@link Pointer#Pointer(Pointer)}. */
     public DoubleCuDNNLCNLayer(Pointer p) { super(p); }
 
   public DoubleCuDNNLCNLayer(@Const @ByRef LayerParameter param) { super((Pointer)null); allocate(param); }
-  private native @Platform({"linux-x86_64", "macosx-x86_64"}) void allocate(@Const @ByRef LayerParameter param);
+  private native @Platform(value = {"linux-x86_64", "macosx-x86_64"}, extension = "-gpu") void allocate(@Const @ByRef LayerParameter param);
   @Virtual public native void LayerSetUp(@Const @ByRef DoubleBlobVector bottom,
         @Const @ByRef DoubleBlobVector top);
   @Virtual public native void Reshape(@Const @ByRef DoubleBlobVector bottom,
@@ -26194,13 +26194,13 @@ public static final String HDF5_DATA_LABEL_NAME = "label";
  * @brief cuDNN implementation of PoolingLayer.
  *        Fallback to PoolingLayer for CPU mode.
 */
-@Platform({"linux-x86_64", "macosx-x86_64"}) @Name("caffe::CuDNNPoolingLayer<float>") @NoOffset public static class FloatCuDNNPoolingLayer extends FloatPoolingLayer {
+@Platform(value = {"linux-x86_64", "macosx-x86_64"}, extension = "-gpu") @Name("caffe::CuDNNPoolingLayer<float>") @NoOffset public static class FloatCuDNNPoolingLayer extends FloatPoolingLayer {
     static { Loader.load(); }
     /** Pointer cast constructor. Invokes {@link Pointer#Pointer(Pointer)}. */
     public FloatCuDNNPoolingLayer(Pointer p) { super(p); }
 
   public FloatCuDNNPoolingLayer(@Const @ByRef LayerParameter param) { super((Pointer)null); allocate(param); }
-  private native @Platform({"linux-x86_64", "macosx-x86_64"}) void allocate(@Const @ByRef LayerParameter param);
+  private native @Platform(value = {"linux-x86_64", "macosx-x86_64"}, extension = "-gpu") void allocate(@Const @ByRef LayerParameter param);
   @Virtual public native void LayerSetUp(@Const @ByRef FloatBlobVector bottom,
         @Const @ByRef FloatBlobVector top);
   @Virtual public native void Reshape(@Const @ByRef FloatBlobVector bottom,
@@ -26213,13 +26213,13 @@ public static final String HDF5_DATA_LABEL_NAME = "label";
   @Virtual protected native void Backward_gpu(@Const @ByRef FloatBlobVector top,
         @Const @ByRef BoolVector propagate_down, @Const @ByRef FloatBlobVector bottom);
 }
-@Platform({"linux-x86_64", "macosx-x86_64"}) @Name("caffe::CuDNNPoolingLayer<double>") @NoOffset public static class DoubleCuDNNPoolingLayer extends DoublePoolingLayer {
+@Platform(value = {"linux-x86_64", "macosx-x86_64"}, extension = "-gpu") @Name("caffe::CuDNNPoolingLayer<double>") @NoOffset public static class DoubleCuDNNPoolingLayer extends DoublePoolingLayer {
     static { Loader.load(); }
     /** Pointer cast constructor. Invokes {@link Pointer#Pointer(Pointer)}. */
     public DoubleCuDNNPoolingLayer(Pointer p) { super(p); }
 
   public DoubleCuDNNPoolingLayer(@Const @ByRef LayerParameter param) { super((Pointer)null); allocate(param); }
-  private native @Platform({"linux-x86_64", "macosx-x86_64"}) void allocate(@Const @ByRef LayerParameter param);
+  private native @Platform(value = {"linux-x86_64", "macosx-x86_64"}, extension = "-gpu") void allocate(@Const @ByRef LayerParameter param);
   @Virtual public native void LayerSetUp(@Const @ByRef DoubleBlobVector bottom,
         @Const @ByRef DoubleBlobVector top);
   @Virtual public native void Reshape(@Const @ByRef DoubleBlobVector bottom,