+++ caffe/src/main/java/org/bytedeco/javacpp/presets/caffe.java
@@ -58,7 +58,7 @@ import org.bytedeco.javacpp.tools.InfoMapper;
         "caffe/util/db_leveldb.hpp", "caffe/util/db_lmdb.hpp", "caffe/util/io.hpp", "caffe/util/rng.hpp", "caffe/util/im2col.hpp", "caffe/util/insert_splits.hpp", "caffe/util/mkl_alternate.hpp",
         "caffe/util/upgrade_proto.hpp", /* "caffe/util/cudnn.hpp" */}, link = "caffe@.1.0.0", resource = {"include", "lib"}, includepath = {"/usr/local/cuda/include/",
         "/System/Library/Frameworks/vecLib.framework/", "/System/Library/Frameworks/Accelerate.framework/"}, linkpath = "/usr/local/cuda/lib/"),
-    @Platform(value = {"linux-x86_64", "macosx-x86_64"}, define = {"SHARED_PTR_NAMESPACE boost", "USE_LEVELDB", "USE_LMDB", "USE_OPENCV", "USE_CUDNN"}) })
+    @Platform(value = {"linux-x86_64", "macosx-x86_64"}, define = {"SHARED_PTR_NAMESPACE boost", "USE_LEVELDB", "USE_LMDB", "USE_OPENCV", "USE_CUDNN"}, extension = "-gpu") })
 public class caffe implements InfoMapper {
     public void map(InfoMap infoMap) {
         infoMap.put(new Info("LIBPROTOBUF_EXPORT", "LIBPROTOC_EXPORT", "GOOGLE_PROTOBUF_VERIFY_VERSION", "GOOGLE_ATTRIBUTE_ALWAYS_INLINE", "GOOGLE_ATTRIBUTE_DEPRECATED",
@@ -127,7 +127,7 @@ public class caffe implements InfoMapper {
         for (String t : classTemplates) {
             boolean purify = t.equals("BaseDataLayer") || t.equals("LossLayer") || t.equals("NeuronLayer");
             boolean virtualize = t.endsWith("Layer") || t.endsWith("Solver");
-            String[] annotations = t.startsWith("CuDNN") ? new String[] {"@Platform({\"linux-x86_64\", \"macosx-x86_64\"})"} : null;
+            String[] annotations = t.startsWith("CuDNN") ? new String[] {"@Platform(value = {\"linux-x86_64\", \"macosx-x86_64\"}, extension = \"-gpu\")"} : null;
             infoMap.put(new Info("caffe::" + t + "<float>").annotations(annotations).pointerTypes("Float" + t).purify(purify).virtualize(virtualize))
                    .put(new Info("caffe::" + t + "<double>").annotations(annotations).pointerTypes("Double" + t).purify(purify).virtualize(virtualize));
         }