+++ tensorflow/src/main/java/org/bytedeco/javacpp/presets/tensorflow.java
@@ -148,6 +148,7 @@ import java.lang.annotation.Target;
                         "tensorflow/cc/ops/training_ops.h",
                         "tensorflow/cc/ops/user_ops.h"},
                 link = {"tensorflow_framework", "tensorflow_cc"}),
+        @Platform(value = {"linux-x86_64", "macosx-x86_64"}, extension = "-gpu"),
         @Platform(value = {"android"},
                 compiler = {"cpp11"},
                 define = {"NDEBUG", "UNIQUE_PTR_NAMESPACE std"},